requirements.txt
# ============================================================
# AI-Powered Medical Image Analysis System
# Requirements File
# ============================================================

# Core Deep Learning
tensorflow==2.13.0
keras==2.13.1

# Image Processing
opencv-python==4.8.0.76
Pillow==10.0.1
scikit-image==0.21.0

# Data Science
numpy==1.24.3
pandas==2.0.3
scipy==1.11.3

# Machine Learning Utilities
scikit-learn==1.3.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# Explainability (Grad-CAM)
tf-keras-vis==0.8.5

# Web App - Streamlit UI
streamlit==1.27.0
streamlit-option-menu==0.3.6

# API Layer
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6

# Utilities
tqdm==4.66.1
requests==2.31.0
pyyaml==6.0.1
python-dotenv==1.0.0

# Jupyter (for notebooks)
jupyter==1.0.0
ipykernel==6.25.2
numpy
tensorflow
matplotlib
scikit-learn
pillow==9.5.0
pyyaml
streamlit